big late commit with para taken from para.h and put into source code

+++ include/para.h
@@ -1,31 +1,31 @@
 #include <vector>
 
-extern std::vector <std::vector <float> > para12;
+extern std::vector <std::vector <double> > para12;
 
-extern  std::vector <std::vector <float> > para15;
+extern  std::vector <std::vector <double> > para15;
 
-extern std::vector <std::vector <float> > para18;
+extern std::vector <std::vector <double> > para18;
 
-extern std::vector <std::vector <float> > para20;
+extern std::vector <std::vector <double> > para20;
 
-extern std::vector <std::vector <float> > para24;
+extern std::vector <std::vector <double> > para24;
 
-extern std::vector <std::vector <float> > para30;
+extern std::vector <std::vector <double> > para30;
 
-extern std::vector <std::vector <float> > para36;
+extern std::vector <std::vector <double> > para36;
 
-extern std::vector <std::vector <float> > para40;
+extern std::vector <std::vector <double> > para40;
 
-extern std::vector <std::vector <float> > para45;
+extern std::vector <std::vector <double> > para45;
 
-extern std::vector <std::vector <float> > para60;
+extern std::vector <std::vector <double> > para60;
 
-extern std::vector <std::vector <float> > para72;
+extern std::vector <std::vector <double> > para72;
 
-extern std::vector <std::vector <float> > para90;
+extern std::vector <std::vector <double> > para90;
 
-extern std::vector <std::vector <float> > para120;
+extern std::vector <std::vector <double> > para120;
 
-extern std::vector <std::vector <float> > para180;
+extern std::vector <std::vector <double> > para180;
 
-extern std::vector <std::vector <float> > para360;
+extern std::vector <std::vector <double> > para360;
